+++ packages/react-headless/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @opencx/widget-react-headless
 
+## 4.0.13
+
+### Patch Changes
+
+- add `WidgetConfig.specialComponents`
+- Updated dependencies
+  - @opencx/widget-core@4.0.13
+
 ## 4.0.12
 
 ### Patch Changes